backend/requirements.txt
# HealthVoice Backend Requirements
# Python Flask backend with MongoDB support

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Database
pymongo==4.6.1

# Environment Variables
python-dotenv==1.0.0

# PDF Generation
reportlab==4.0.9

# Date/Time Utilities (included in standard library, but listed for reference)
# datetime - built-in

# Text Processing (included in standard library)
# re - built-in
# collections - built-in

